Endrer i mappestruktur
