+++ src/søknad/steg/5-arbeidssituasjon/underUtdanning/ErUtdanningenOffentligEllerPrivat.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import MultiSvarSpørsmål from '../../../../../components/spørsmål/MultiSvarSpørsmål';
+import MultiSvarSpørsmål from '../../../../components/spørsmål/MultiSvarSpørsmål';
 import { privatEllerOffentligSpm } from './UtdanningConfig';
-import KomponentGruppe from '../../../../../components/gruppe/KomponentGruppe';
-import { IUnderUtdanning } from '../../../../../models/arbeidssituasjon/utdanning';
-import { ISpørsmål } from '../../../../../models/spørsmal';
-import { hentTekst } from '../../../../../utils/søknad';
+import KomponentGruppe from '../../../../components/gruppe/KomponentGruppe';
+import { IUnderUtdanning } from '../../../../models/arbeidssituasjon/utdanning';
+import { ISpørsmål } from '../../../../models/spørsmal';
+import { hentTekst } from '../../../../utils/søknad';
 import { useIntl } from 'react-intl';
 interface Props {
   utdanning: IUnderUtdanning;
